Condense Crosses to one file
Switch Hero and Crosses to tailwind
Comment out nav until app routing gets migrated
