--- src/app/_pagecomponents_/Hero/Crosses/Crosses.module.scss
@@ -1,9 +1,0 @@
-.container {
-	position: absolute;
-	height: 100svh;
-	width: 100%;
-	overflow-x: hidden;
-	overflow-y: hidden;
-	perspective: 400px;
-	z-index: 1;
-}
