--- src/app/_pagecomponents_/Hero/Crosses/Crosses.tsx
@@ -1,27 +1,0 @@
-'use client';
-
-import { useMemo } from 'react';
-
-import Cross from './Cross/Cross';
-
-import crossesLook from './Crosses.module.scss';
-
-const Crosses = ({ limit }: { limit: number }) => {
-	const initer = useMemo(() => {
-		let arr = [];
-		for (let i = 0; i < limit; i++) {
-			arr.push(i);
-		}
-		return arr;
-	}, [limit]);
-
-	return (
-		<div className={crossesLook.container}>
-			{initer.map((slot, index) => {
-				return <Cross key={`cross-${index}`} />;
-			})}
-		</div>
-	);
-};
-
-export default Crosses;
