--- src/app/_pagecomponents_/Hero/Hero.module.scss
@@ -1,43 +1,0 @@
-@use '@styles/globalStyles.module.scss' as global;
-
-.hOne {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	width: 100%;
-	height: 100svh;
-	z-index: 2;
-}
-
-.hOne * {
-	margin: 0;
-}
-
-.name {
-	font-family: Poppins, sans-serif;
-	font-size: 8vmax;
-	font-weight: bold;
-}
-
-.title {
-	font-family: Lato, sans-serif;
-	font-size: 4vmax;
-	font-weight: lighter;
-}
-
-@media (min-aspect-ratio: 1) {
-	.name {
-		font-weight: bold;
-	}
-
-	.title {
-		font-weight: lighter;
-	}
-}
-
-.line {
-	width: 25%;
-	opacity: 0.5;
-
-}
