+++ src/app/_pagecomponents_/Hero/Hero.tsx
@@ -1,15 +1,19 @@
-import Crosses from './Crosses/Crosses';
+import Crosses from './Crosses';
 
-import heroLook from './Hero.module.scss';
+import { poppinsClass, latoClass } from '@/styles/fonts';
 
 const Hero = () => {
 	return (
-		<header className={heroLook.hOne}>
-			<h1 className={heroLook.name}>Nicolas Vo</h1>
-			<h2 className={heroLook.title}>Front End Developer</h2>
+		<header className='flex flex-col items-center justify-center w-full h-svh z-10'>
+			<h1 className={poppinsClass + ' text-[8vmax] font-bold m-0'}>
+				Nicolas Vo
+			</h1>
+			<h2 className={latoClass + ' text-[4vmax] font-light m-0'}>
+				Front End Developer
+			</h2>
 			<Crosses limit={32} />
 		</header>
 	);
-}
+};
 
 export default Hero;
